lpw_04/requirements.txt
pandas==2.3.3
kaggle==1.5.16
psycopg2-binary==2.9.7
apache-airflow-providers-postgres==5.7.0
sqlalchemy==2.0.23
